+++ javascripts/components/river.js
@@ -1,12 +1,13 @@
 
-import addBear from './bear.js';
-import printToDom from '../helpers/data/util.js';
+import addBear from './bears.js';
+import util from '../helpers/data/util.js';
+import bearData from '../helpers/data/bearData.js';
 
 //create card for bear, need click event
 const cardBears = ()=> {
+    const bearDom = bearData.getBears();
     let domstring="";
-    console.log("mememe");{
-    allBears.forEach(bears);
+        bearDom.forEach((bears) => {
         domstring += '<div class = "d flex flex-wrap">';
         domstring += '<div class="card" style="width: 18rem;">';
         domstring += `<img src="${bears.image}" class="card-img-top" alt="...">`;
@@ -15,10 +16,11 @@ const cardBears = ()=> {
         domstring += '</div>';
         domstring += '</div>';
 
-}
+        })
+    
 
-util.printToDom('river', domstring);
-};
+    util.printToDom('river', domstring);
+}
 
 
 export default { cardBears };
