+++ javascripts/helpers/data/util.js
@@ -1,6 +1,8 @@
+
+//print to dom 
 const printToDom = (divId, textToPrint) => {
     const selectedDiv = document.getElementById(divId);
     selectedDiv.innerHTML = textToPrint;
   };
   
-  export default { printToDom };
+export default { printToDom };
