--- main.js
@@ -1,14 +1,0 @@
-import bearEvent from './javascripts/components/bear.js';
-import printToDom from './javascripts/helpers/data/util.js';
-import cardBears from './javascripts/components/river.js';
-import bearData from './javascripts/helpers/data/bearData.js';
-
-const init = () =>{
-
-bear.addBearEvent();
-bearData.getBears();
-
-};
-
-
-init();
